Move tests to their own file and tagging them  to not run on the FPGA.
