+++ rom/dev/tests/rom_integration_tests/main.rs
@@ -9,6 +9,7 @@ mod test_cm_sha;
 mod test_cpu_fault;
 mod test_debug_unlock;
 mod test_derive_stable_key;
+mod test_device_status_mode;
 mod test_dice_derivations;
 mod test_ecdsa_verify;
 mod test_fake_rom;
